[refactor] 감상 상세보기 / ImageBindingAdapter로 합치기 (#119)
