+++ android/app/src/main/res/layout/item_post_viewer.xml
@@ -31,7 +31,7 @@
             app:layout_constraintEnd_toStartOf="@id/tv_post_title"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="@id/tv_post_title"
-            app:setThumbnail="@{postItem.thumbnail}"
+            app:setRoundImage="@{postItem.thumbnail}"
             tools:background="@drawable/shape_td_gray_fill_12_rect" />
 
         <TextView
